Convenience class method on NSBundle for localized strings
It looks like we don't have access to the commonly used macro,
NSLocalizedString (defined in NSBundle.h).

I saw that James Chen addressed this by declaring an NSlocalizedString
function in rb_main:

  https://github.com/ashchan/gmail-notifr/blob/master/rb_main.rb#L25

This is surely the most readable solution for veteran Cocoa programmers
used to seeing the macro peppered throughout code, I wonder if it
belongs within the global namespace, and if there is a better macro
mapping solution.